refactor: rename model files
