+++ server/src/models/agreement.model.ts
