+++ server/src/models/challenge.model.ts
