+++ server/src/models/item.model.ts
